refactor: 테스트 코드를 calculator 패키지로 옮김
