+++ src/test/java/study/calculator/StringAdditionTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
